Extend rollout script (#29)
Adds some switches to control the behavior of the rollout script:

-p: Push is only done if -p is set, otherwise it is just a dry run
-f: Override existing gradle settings
-u: Only update existing project with .mvn Folders

+++ maven-extension-rollout/rollout.sh
@@ -2,18 +2,27 @@
 
 commit_msg="Apply Gradle Enterprise Maven Extension"
 basedir="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
-repositories=$basedir/repositories.txt
-extensions_xml=$basedir/extensions.xml
 checkout_area=`mktemp -d`
+repositories=$basedir/repositories.txt
+
+# process arguments "$1", "$2", ... (i.e. "$@")
+while getopts "pfu" opt; do
+    case $opt in
+    p) push=true ;;   # Push only if -p is set
+    f) force=true ;;  # Override existing gradle settings
+    u) do_update=true ;; # Only update existing project with .mvn Folders 
+    \?) ;; # Handle error: unknown option or missing required argument.
+    esac
+done
 
 function process_repositories() {
-  num=`wc -l "$repositories" | awk '{ print $1 }'`
+  num=$(cat $repositories | sed '/^\s*$/d' | wc -l | awk '{ print $1 }')
   i=1
-  while read line; do
-  echo "($i/$num) Updating $line..."
-  process_repository $line
-  ((i++))
-  done < $repositories
+  for line in $(cat repositories.txt | grep .); do
+    echo "($i/$num) Updating $line..."
+    process_repository $line
+    ((i++))
+  done
 }
 
 function process_repository() {
@@ -22,37 +31,49 @@ function process_repository() {
   git clone -n "$1" "$checkout_area/$repository_name" --depth 1 >& /dev/null
   pushd "$checkout_area/$repository_name" >& /dev/null
   git reset HEAD . >& /dev/null
-  git checkout .mvn/extensions.xml >& /dev/null
-  if [ $? -eq 0 ]; then
-    echo ".mvn/extensions.xml already exists in $1, skipping..." >&2
-  else
+  if [ ! -z  "$(git ls-tree -r HEAD --name-only | grep '^.mvn')" ] && [ -z "$force" ] && [ -z "$do_update" ]; then
+    echo ".mvn directory already exists in $repository_name, skipping..." >&2
+  elif [ -z "$(git ls-tree -r HEAD --name-only | grep pom.xml)" ]; then  
     # Only process maven projects
-    if [ -z "$(git ls-tree -r HEAD --name-only | grep pom.xml)" ]; then  
-      echo "$1 is not a mvn project, skipping..." >&2
-    else 
-      mkdir -p .mvn
-      cp "$extensions_xml" .mvn
-      git add .mvn/extensions.xml >& /dev/null
-      git checkout .gitignore >& /dev/null
-      echo ".mvn/.gradle-enterprise/" >> .gitignore
-      git add .gitignore
-      git commit -m "$commit_msg" >& /dev/null
+    echo "$repository_name is not a mvn project, skipping..." >&2
+  elif [ -z  "$(git ls-tree -r HEAD --name-only | grep '^.mvn')" ] && [ ! -z "$do_update" ]; then
+    echo "$repository_name is not enabled for gradle caching, skipping..." >&2
+  else
+    cp -R $basedir/.mvn/ .mvn
+    
+    git add .mvn/. >& /dev/null
+    insert_gitignore
+    git commit -m "$commit_msg" >& /dev/null
 
+    if [ ! -z "$push" ]; then
       git push >& /dev/null
     fi
   fi
   popd >& /dev/null
 }
 
-function cleanup() {
-  rm -rf $checkout_area
+function insert_gitignore() {
+  git checkout -- .gitignore >& /dev/null
+  if [ $? -eq 0 ]; then
+    if ! grep -Fxq "/.mvn/.gradle-enterprise/" .gitignore ; then
+      echo "/.mvn/.gradle-enterprise/" >>  .gitignore
+    fi
+  else
+    echo "/.mvn/.gradle-enterprise/" > .gitignore
+  fi
+  git add .gitignore >& /dev/null
 }
 
-if [ ! -f "$repositories" ]; then
-  echo "File $repositories is missing" >&2
-  exit 1
-elif [ ! -f "$extensions_xml" ]; then
-  echo "File $extensions_xml is missing" >&2
+function cleanup() {
+  if [ ! -z "$push" ]; then
+    rm -rf $checkout_area
+  else
+    echo "Stored repos to " $checkout_area
+  fi
+}
+
+if [ ! -d ".mvn" ]; then
+  echo ".mvn directory is missing" >&2
   exit 1
 else
   process_repositories
